chore(refactor): move models to relevant redux modules
