+++ src/common/redux/types.ts
@@ -1,18 +1,22 @@
 import { KeysafeState } from '../modules/keysafe/types'
 import { LoginState } from '../../modules/login/types'
 import { IxoState } from '../../common/modules/ixo/types'
+import { QuoteState } from '../../modules/quote/types'
+import { BondState } from '../../modules/bond/types'
+import { Currency } from '../../types/models'
+import { AccountState } from '../../modules/account/types'
 
 export interface RootState {
   keySafe: KeysafeState
   ixo: IxoState
   login: LoginState
   web3: any
-  activeQuote: any
-  quotePending: any
-  signPending: any
-  transacting: any
-  account: any
-  balances: any
-  activeBond: any
-  totalSupplies: any
+  activeQuote: QuoteState
+  quotePending: boolean
+  signPending: boolean
+  transacting: boolean
+  account: AccountState
+  balances: Currency[]
+  activeBond: BondState
+  totalSupplies: Currency[]
 }
