+++ src/components/bonds/exchange/buy/ConfirmBuy.tsx
@@ -2,14 +2,14 @@ import React from 'react'
 import useForm from 'react-hook-form'
 import { withRouter, Redirect } from 'react-router-dom'
 import { connect } from 'react-redux'
-import { Store } from '../../../../model/store'
+import { RootState } from '../../../../common/redux/types'
 
 import { confirmBuy, clearQuote } from '../../../../modules/quote/quote.actions'
 import {
   remainingBalance,
   newBalance,
   currencyStr,
-} from '../../../../model/account'
+} from '../../../../modules/account/account.utils'
 
 const ConfirmBuy = (props: any): JSX.Element => {
   const { handleSubmit } = useForm()
@@ -139,7 +139,7 @@ const ConfirmBuy = (props: any): JSX.Element => {
   }
 }
 
-const mapStateToProps = (state: Store): Store => {
+const mapStateToProps = (state: RootState): RootState => {
   return state
 }
