+++ src/components/bonds/exchange/sell/ConfirmSell.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import useForm from 'react-hook-form'
 import { withRouter, Redirect } from 'react-router-dom'
 import { connect } from 'react-redux'
-import { Store } from '../../../../model/store'
+import { RootState } from '../../../../common/redux/types'
 
 import {
   confirmSell,
@@ -13,7 +13,7 @@ import {
   remainingBalance,
   newBalance,
   currencyStr,
-} from '../../../../model/account'
+} from '../../../../modules/account/account.utils'
 
 const ConfirmSell = (props: any): JSX.Element => {
   const { handleSubmit } = useForm()
@@ -130,7 +130,7 @@ const ConfirmSell = (props: any): JSX.Element => {
   }
 }
 
-const mapStateToProps = (state: Store): Store => {
+const mapStateToProps = (state: RootState): RootState => {
   return state
 }
