+++ src/components/bonds/exchange/sell/QuoteSell.tsx
@@ -2,16 +2,19 @@ import React from 'react'
 import useForm from 'react-hook-form'
 import { withRouter } from 'react-router-dom'
 import { connect } from 'react-redux'
-import { Store } from '../../../../model/store'
+import { RootState } from '../../../../common/redux/types'
 import { quoteSell } from '../../../../modules/quote/quote.actions'
-import { Quote } from '../../../../model/quote'
-import { currencyStr, tokenBalance } from '../../../../model/account'
+import { QuoteState } from '../../../../modules/quote/types'
+import {
+  currencyStr,
+  tokenBalance,
+} from '../../../../modules/account/account.utils'
 
 const QuoteSell = (props: any): JSX.Element => {
   const { register, handleSubmit, watch, errors } = useForm()
 
   const onSubmit = (formData: any): void => {
-    const quote: Quote = {}
+    const quote: QuoteState = {}
     quote.recieving = { denom: formData.denom }
     quote.sending = { amount: formData.amount, denom: props.activeBond.symbol }
     quote.minPrices = [{ denom: formData.denom, amount: formData.minAmount }]
@@ -113,7 +116,7 @@ const QuoteSell = (props: any): JSX.Element => {
   }
 }
 
-const mapStateToProps = function(state: Store): Store {
+const mapStateToProps = function(state: RootState): RootState {
   return state
 }
