+++ src/components/bonds/orders/Orders.tsx
@@ -1,7 +1,7 @@
 import React, { Component } from 'react'
 import 'react-virtualized/styles.css'
 import { connect } from 'react-redux'
-import { Store } from '../../../model/store'
+import { RootState } from '../../../common/redux/types'
 import { getOrders } from '../../../modules/account/account.actions'
 import TransactionsTable from '../TransactionsTable'
 import BondsWrapper from '../BondsWrapper'
@@ -31,7 +31,7 @@ class Orders extends Component<any> {
   }
 }
 
-const mapStateToProps = function(state: Store): Store {
+const mapStateToProps = function(state: RootState): RootState {
   return state
 }
