+++ src/modules/quote/quote.actions.ts
@@ -1,9 +1,9 @@
-import { QuoteAction, Quote, QuoteActions } from '../../model/quote'
+import { QuoteAction, QuoteState, QuoteActions } from './types'
 import { AsyncAction } from 'redux-promise-middleware'
 import Axios from 'axios'
-import { Currency } from '../../model'
-import { currencyStr } from '../../model/account'
-import { Bond } from '../../model/bond'
+import { Currency } from '../../types/models'
+import { currencyStr } from '../account/account.utils'
+import { BondState } from '../bond/types'
 import KeystationService from '../../service/KeystationService'
 import { toast } from 'react-toastify'
 const ks = new KeystationService()
@@ -16,7 +16,7 @@ export function clearQuote(): QuoteAction {
 
 // BUYING
 
-export const quoteBuy = (quote: Quote): AsyncAction => {
+export const quoteBuy = (quote: QuoteState): AsyncAction => {
   return {
     type: QuoteActions.QUOTE_BUY,
     payload: Axios.get(
@@ -32,8 +32,8 @@ export const quoteBuy = (quote: Quote): AsyncAction => {
 }
 
 export const confirmBuy = (
-  quote: Quote,
-  bond: Bond,
+  quote: QuoteState,
+  bond: BondState,
   address: string,
 ): AsyncAction => {
   const maxPrices = quote
@@ -79,7 +79,7 @@ export const confirmBuy = (
 
 // SELLING
 
-export const quoteSell = (quote: Quote): AsyncAction => {
+export const quoteSell = (quote: QuoteState): AsyncAction => {
   return {
     type: QuoteActions.QUOTE_SELL,
     payload: Axios.get(
@@ -95,8 +95,8 @@ export const quoteSell = (quote: Quote): AsyncAction => {
 }
 
 export const confirmSell = (
-  quote: Quote,
-  bond: Bond,
+  quote: QuoteState,
+  bond: BondState,
   address: string,
 ): AsyncAction => {
   const minPrices = quote
@@ -144,7 +144,7 @@ export const confirmSell = (
 
 // SWAPPING
 
-export const quoteSwap = (quote: Quote): AsyncAction => {
+export const quoteSwap = (quote: QuoteState): AsyncAction => {
   return {
     type: QuoteActions.QUOTE_SWAP,
     payload: Axios.get(
@@ -162,8 +162,8 @@ export const quoteSwap = (quote: Quote): AsyncAction => {
 }
 
 export const confirmSwap = (
-  quote: Quote,
-  bond: Bond,
+  quote: QuoteState,
+  bond: BondState,
   address: string,
 ): AsyncAction => {
   const payload = {
